requirements.txt
auto-py-to-exe==2.23.1
pillow==9.3.0
